requirements.txt
flask
flask_restful
flask_pymongo
mongomock
bson
bcrypt
pyjwt
redis